move maya_sequencer file
